user: Don't delete empty files
Previously we deleted the destination file if no progress was made
during an upload/download. This is especially convenient if an
upload fails because not enough hosts are reachable, because you
can just try again with different hosts (if the .usa file *isn't*
deleted, you have to migrate it or delete it manually before
"resuming" the upload with different hosts). But this behavior
breaks uploading/downloading files that are intentionally empty.

Probably the Right Thing To Do is to delete the .usa file IFF it
has no shard data AND m.Filesize > 0. For now, do the simple thing
and never delete.