proto: Delay Revise until we have host signature
Previously, we wrote the revision to the contract file immediately
prior to proposing the revision to the host. The problem with this
is that the revision won't have the host's signatures, so when you
try to broadcast the revision transaction (with 'tattle'), the
transaction will be in an invalid state. This commit moves the
Revise call after negotiation, so that we can write it with the
new host signature.

To be honest, I'm not entirely sure why we were writing the
unsigned revision. I mean, there *is* a good reason to append
the new sector Merkle root: it ensures that if something goes
wrong during negotiation (e.g. a power-loss event), you can
always recover to either the pre- or post-revision state. But
the same logic doesn't apply to the revision itself, since you
can always get that from the host. Indeed, this is what enables
the Revise call to overwrites the revision part of the .contract
file in an unsafe manner: if the write is interrupted, corrupting
the revision, we're still okay as long as we have the renter key,
contract ID, and Merkle roots.

So I'm fairly sure this is safe, but not 100%. I'll do some manual
testing to see if I can identify any edge-cases.