all: ContractTransaction -> ContractRevision
This is a natural continuation of the previous commit. Note that
instead of storing a types.Transaction, the .contract format now
stores the ContractRevision, including the RenterKey field. So now
in LoadContract we check that both the FileContractID and RenterKey
of the ContractRevision match the .contract header. You might wonder
why we don't simply overwrite these fields with the (known correct)
values in the header. After all, this is pretty much always what the
user would want. But on the off chance that something really weird
happened, I don't want to destroy the evidence without at least
giving the user a chance to inspect it and (hopefully) report it.
I guess that means I need to add a user command for fixing .contract
files with corrupted revisions...