renter: Use MerkleStack in Contract
This dramatically reduces memory usage when using large contracts.
For example, 50 contracts each storing 1 TiB of data consume a total
of 50 * (1 TiB / 4 MiB) * 32 bytes = 400 MiB. On disk, they still
consume that much space, but now in memory they only consume at most
50 * 64 * 32 bytes = 102.4 KB. (The stack implementation uses a
static 64-element array instead of a dynamic one, so it's not quite
as efficient as log2, but it's easier to work with.)

This means that we can basically ignore the impact of contracts on
total memory usage. I haven't profiled it yet, but I'm pretty sure
the only big memory consumer left is the erasure encoding. With 50
hosts, you need 50 * 4 MiB = 200 MiB to encode everything. That's
not terrible; you could run dozens of instances on today's hardware.
But if we really wanted to, we could get this down even further by
doing the encoding in multiple passes (e.g. encode the first 10
shards, then the next 10, etc.). There's a performance penalty for
this: you can't transfer every shard in parallel anymore. But that's
likely not an issue unless you have a really fast connection. Even
1 Gbps only gets you 30 sectors per second!