renter: Fix bufio breaking SyncWithHost
This is an interesting bug that happened because I wasn't careful
enough when doing buffered I/O. SyncWithHost reads from c.f in two
places, but in the first one it uses a bufio.Reader and in the other
it doesn't. As a result, there were bytes left over in the buffer
that were silently dropped, causing subsequent reads to hit EOF. And
ReadFrom drops io.EOF errors (as per the interface spec), it became
a no-op and the loop never terminated.

The fix is simply to move the bufio outside the limit reader. This
means we can still do buffered I/O, but we'll never read more bytes
than we should.