proto: Add (Build|Verify)MerkleProof
These aren't useful right now, but they will be once the host starts
using them to provide proofs that arbitrary segments are part of the
full tree. This would let us completely eliminate local checksums!
(Although for performance, some users will still want them.)

Indeed, the main reason why I wrote these functions now is so that
I could benchmark them and thus determine how feasible it is for a
host to provide such proofs. The results are fairly good: building
a proof takes ~20ms, which isn't amazing, but this can be greatly
improved by storing a few extra subtree roots. Moreover, the size
of the proof is optimal: it's just a slice of hashes, no extra
metadata required. This is possible because both the prover and
the verifier know what the initial conditions are, so they can
traverse the tree in a deterministic manner.

If we actually deploy this, we'll definitely need a way for the host
to store subtree roots and tell BuildMerkleProof about them somehow.
I haven't devoted much thought to this yet, but I expect it will be
a fun challenge.