proto: Allow precalculated subtree roots
Benchmarks indicate that this does result in the predicted speedup:
providing just one precalculated root for one half of the tree speeds
up BuildMerkleProof by about 2x. In the optimal case (all the roots
relevant to the proof are precalculated), BuildMerkleProof doesn't do
any hashing at all; it runs in 500ns! I don't know how the economics
of this will work out, but it would be pretty cool if hosts started
storing tons of roots to save on hashing time. On the other hand,
going to disk is a lot more expensive than computing a few thousand
hashes, so unless they can amortize that I/O somehow, in practice
they'll likely only store roots to depth 3 or 4.

The API for this isn't great; the precalc functions end up looking
ugly. It would be a bit nicer if the arguments were (depth, index) or
something, because that's how you'd typically think of caching (cache
at depth 1, then at depth 2, then 3 if that's not fast enough, etc.)
but it's good enough for now. Once I know what the host code for this
will actually look like, it'll be easier to figure out a good API.