proto: Use MerkleStack in VerifyMerkleProof
Reusing the s.merge method is slightly ugly, but the results are
so good that I don't care: VerifyMerkleProof now allocates zero
memory! This is particularly nice for large proof ranges, since
the old code was naively allocating a slice to hold all of the
leaf hashes. We still have to calculate them, but there's no need
to compute them all at the beginning and store them in a big slice.