renter: Switch encryption to ChaCha20
I got excited about XTS for a few reasons. First, it has zero
overhead, which immediately makes it nicer than AEAD. Second,
because it's a tweakable cipher, it supports edits: unlike CTR
mode, you can re-encrypt different data with the "same key"
without leaking key material. Perfect, right?

Well, sort of. XTS has a major downside: it's slow. Even when
using AES, XTS-AES is at least an order of magnitude slower than
bare AES. I think the main reason for this is actually the fault
of the xts package I used: from what I understand, XTS shouldn't
add much overhead. I'm guessing the reason is that the code has
to frequently jump between handwritten asm and normal Go code.
The AES part is fast, but profiling shows that the benchmark only
spends about 40% of its time there; the rest is spent xor'ing and
updating the tweak. I'd really like to see what the performance
would look like if the whole thing were rewritten in asm.

So XTS is slow, but it supports edits, so doesn't that make up
for it? Not really, because they're the *only* part of the system
that supports edits! Editing a metafile is a hard problem, mostly
because editing a sector changes its Merkle root, so you need to
run around and fix all the metafiles that reference the old root.
Maybe that's possible when all your metafiles are local, but once
you start sharing them...you see the problem. So why should we
bother supporting a feature that may not ever materialize? Let's
focus on optimizing the functionality we have now.

Which brings us to ChaCha20. It's crazy fast, almost as fast as
AES, without the grossness of AES. Actually, when AES is used in
CTR mode, its speed drops significantly, so ChaCha20 is much faster!
(This is probably because ChaCha20 was designed as a stream cipher
from the beginning, whereas AES is a block cipher that has to be
wrapped in extra Go code to make it a stream cipher.) ChaCha20 is
made by djb so we know it's good. It also happens to be the basis
of Blake2b. Functionally, it brings us all the advantages of XTS
(zero overhead, random access), sacrificing only the edit safety.
What's not to love?

Sure, there might be some benefit to "future-proofing" metafiles,
such that we don't need to re-encrypt them if we do ever support
edits. But that's not my style. If we decide that XTS really is
necessary later, we can do that in metafile v2 (or whatever). And
as this diff shows, the amount of work required to switch is pretty
trivial. But for now, we really only care about zero overhead and
random access, and ChaCha20 is the best cipher that supports those
features. Being super fast doesn't hurt either.

My only reservation with ChaCha20 is that I don't know how easy
it is to use in other languages. libsodium supports it, and there
are libsodium bindings for just about every major language, so that's
probably good enough. I'm glad that Go has so many native native
crypto packages. :)