proto: Use smaller buffer for partial downloads
The old code always allocated a full sector, even if you were only
downloading a 1KB file. And the approach of using a separate type
with an UnmarshalSia method was always kind of iffy.

The downside of the new approach is that it will allocate excessively
if you download with length=1, then length=2, etc. We could prevent
that by growing more than strictly necessary, but is it worth it? In
practice I expect that sort of scenario to be extremely rare. Even if
an attacker forces you to execute the worst-case scenario, it's not a
huge deal: one extra allocation per download request. So I'm pretty
sure this is fine.