renter: Add MerkleStack to contract format
Also switch the encoding of the contract revision to binary; this
saves a lot of precious space, now that we need to fit both the
revision and the Merkle stack in 4096 bytes.

The performance impact of this is really dramatic: LoadContract now
runs in more-or-less constant time, since it only ever has to do one
4 KiB read; for a 1 TB contract file (8 MiB of roots), LoadContract
is *three orders of magnitude faster* (100us vs 100ms)! This is a
really important win for the single-invocation model, where you pay
these startup costs over and over again (unlike siad, where you only
pay them once). Thanks to this optimization, we should be able to
load 50 1 TB contracts with almost no noticeable delay, hopefully
less than 10 ms.

I do worry about the long-term usability of the Merkle stack model,
which only supports appends. Contracts just grow and grow forever;
there's no way to delete or modify a sector. Clearly that isn't
sustainable. But that's a problem for another time.