renter: Only fix Merkle root on disk desync
SyncWithHost is now being used to update the revision after every
upload/download, but it was still assuming that the renter's
revision would already be updated. This caused it to unnecessarily
rebuild the MerkleStack on each call, since the revision would never
be updated before SyncWithHost. Instead of checking the renter's
revision, instead just compare the host's Merkle root to diskRoot,
i.e. the root of the MerkleStack on disk. If those match, we can
assume it's safe to just overwrite the renter's revision with the
host's copy.