proto: Clean up payout calculation
It would still be nice to rectify this properly with a hardfork,
but this will do for now. Also, what's the optimal way to converge
on the answer here? Incrementing by 1 obviously works, but I feel
like there must be a Newton's-method-ish algorithm that finds the
answer in O(log(n)).