proto: Have caller provide download buffer
proto is a low-level package so I don't think it should be in charge
of allocating download buffers. This hurts useability somewhat, but
most people will be using DownloadAndDecrypt anyway (which is where
buffer allocation is now performed).

I'm not sure what the ideal strategy for buffer allocation is; it's
tempting to just pre-allocate a [proto.SectorSize]byte, but I don't
like the idea of allocating 200MB to download 1KB from 50 hosts. On
the other hand, allocating dynamically means potentially resizing
many times, which adds GC pressure.

I didn't expect this to improve efficiency, but it turns out that
MigrateDirect can now reuse a single buffer across all hosts, instead
of each Downloader allocating its own buffer. So that's nice.