merkle: Simplify SectorRoot
Previously, SectorRoot was spinning up a bunch of goroutines, with
the intent to minimize the wall time per call. The downside of this
is that it hides concurrency from the caller. A single SectorRoot
call might be fast, but calling it 8 times in parallel would not be
any faster than calling a single-threaded version 8 times in parallel.
Since calling SectorRoot in parallel should be common, I have switched
to a single-threaded version, opting for simplicity over dubious
performance gains. This version is trivially correct, and zero-alloc
too, which is always nice.