merkle: Simplify CachedRoot
Following the footsteps of the SectorRoot change, this commit
simplifies the CachedRoot implementation. Honestly, it's trivial
enough to be written by the caller, but why not provide a little
convenience?

It's really rad that Stacks can be stack-allocated. I don't think
that's possible if you want to support an arbitrary hash.Hash.