all: Add download stats
After much hand-wringing about the design, I finally bit the bullet
and wrote some code. There were basically two options: I could make
Download return the stats on every call, or I could add a getter
method to return the stats of the most recent call. I went with the
latter, mostly because I don't like returning lots of values that the
caller doesn't care about. Is it ideal? No. But the good news is that
I'll have a chance to redesign it once the new RPC protocol is
implemented.

Anyway, this paves the way for user to log useful stats to disk on
each download. I would really like to be able to automatically render
graphs showing exactly when each host started downloading and how
long it took. That makes it dead-simple to identify the slowest hosts
and blacklist in on future downloads, or migrate to new hosts.