proto: Remove spurious TransactionSignature
Not particularly proud of this bug. I guess I never tested the new
SignTransaction API after the most recent change. I was naively
adding the incomplete TransactionSignature at the same point where
the funding inputs were added, not realizing that the host would
end up receiving these TransactionSignatures as well. The fix
simply defers adding them until just before signing.

I guess this demonstrates that the API isn't hard to misuse. This
wouldn't have happened, for example, if SignTransaction filled out
the entire TransactionSignature for you. But this is impossible, so
what are we to do? Maybe force the caller to provide the partial
signatures as arguments, instead of setting them inside the txn?
That doesn't sound too horrible. I'll think about it.