renterutil: Block while sending updates
A consequence of renterutil's channel-based API is that someone needs
to make decisions about whether to buffer the channel, whether to
block while sending or drop missed updates on the floor, etc. Since
renterutil is a *utility* package, that responsibility falls to me,
and my design must assume that the API will be used incompetently.
In the non-blocking design, if the client doesn't process updates
fast enough, they'll miss some. In the blocking design, if they
don't process fast enough, the whole system can time out or just
grind to a halt. Neither option is attractive, but what pushed me
to switch to the latter design is that the new "Stats" updates
arrive so quickly that using an unbuffered channel is no longer
viable; the chance of missing an update is just too high. Sure, I
could switch to using buffered channels, but that just punts the
problem: if you're processing slower than updates are arriving,
eventually the buffer will fill up and you'll start dropping
updates again. So now we block on all sends and hope that the
client doesn't deadlock. Even if they process updates slowly, it
won't cause anything to explode; uploads/downloads will just take
a little longer.