proto: Fix price-increase check false alarm
I recently changed the proto code to not assume that the host had
been scanned prior to creating an Uploader/Downloader. This means
that the initial HostSettings will be empty, and thus the price-
increase check will always trigger if the actual prices are non-zero.
The price-increase check is kind of a hack anyway; I still haven't
hit upon a good design for protecting the renter from the host
jacking up their prices mid-transfer.