user: Log stats from all upload/download cmds
This is kinda indicating to me that the log objects should contain
the name of the file. But that breaks down for packed uploads, so...
it's not clear what the best approach is there. As before, I'm not
going to worry about it too much until someone is actually consuming
these logs.