all: Use 'metafile' everywhere
There were a few instances "meta file" sprinkled around. I'm still
not super happy with the name, but it will have to do for now. Also,
does this mean the MetaFile type should be renamed Metafile? It looks
weird to me that way, but that's probably just due to familiarity.