user: Add support for SHARD servers
A SHARD server is a partially-trusted source of information about
the Sia blockchain. It primarily serves as a renter.HostKeyResolver:
it will relay the most recent host announcement for a given pubkey.
But it also serves the current blockheight, which renters need in
order to calculate upload costs. By providing these two services, a
SHARD server can be used in place of a full siad node for common
user operations (uploading, downloading, migrating). If the user
specifies a SHARD server in their config, it will be used for these
operations; otherwise, siad is used as usual. siad is always used
for forming and renewing contracts.