ed25519: Add internal/ed25519
This is a copy of golang.org/x/crypto/ed25519, plus the performance
improvements from https://go-review.googlesource.com/c/crypto/+/71950

I also stripped out a lot of the code we don't need, and replaced the
Sign/Verify functions with SignHash/VerifyHash methods that satisfy
the interfaces in renterhost. The payoff here is that now we have a
single definition of ed25519 keys, instead of the renter/host/hostdb
all implementing the same tiny wrapper around the native ed25519
package.

A side benefit is that, since hashes are constant-size, SignHash and
VerifyHash incur one less allocation than Sign and Verify. :P