renterutil: Revert multi-slice optimization
In 57e97a20 I split shards into a bunch of 1 MiB SectorSlices before
uploading. I mentioned in that commit message that I had not tested
the change. If I had, I might have discovered that this change was
completely broken, for two reasons. First, it made an incorrect
assumption about how erasure coding works -- namely, that it is
"random access," which it is not. Second, it failed to increment
the chunkIndex properly: it should have been implemented by
(4 MiB / 1 MiB) on each loop iteration, but instead was only
incremented by 1. This meant that the same encryption key was being
used to encrypt multiple 1 MiB chunks -- a big no-no.

I need to refactor the erasure coding before I can restore this
optimization. Also, true partial downloads are just around the
corner, which will have broad effects on file metadata. Specifically,
we currently treat a SectorSlice as the atomic retrievable unit; but
with partial downloads, the granularity drops to 64 bytes. We could
simply use a size of 64 bytes for all SectorSlices, but that would
be prohibitively expensive. So I need to find a more efficient way
of combining the concept of SectorSlices (building a file out of
potentially non-contiguous sector data) with the 64-byte granularity
afforded by partial downloads.