renter: Add Shard tests+benchmarks
As usual, benchmarking revealed that the bottleneck was not what I
expected. I had previously added a buffer to the Shard struct that
would be reused for each Write; but because we're writing directly
to an *os.File, the compiler knows the []byte isn't retained, so
we can just stack-allocate the buffer on each call. In other words,
my premature optimization didn't actually save an allocation!

At the same time, WriteSlice *was* allocating once per call, but
the allocation was coming from the Stat call that checks whether
the file needs to be grown before writing. I replaced the Stat call
with a numSlices field that tracks the size as we go, and the
benchmarks improved significantly. I'm seeing ~75 MB/s write speed
per Shard, up to ~240 MB/s when writing to many Shards in parallel.
And zero allocations, of course. ;)