renter: Overhaul erasure encoding
Basically: Instead of encoding one giant matrix, with one sector
per row, we now encode a bunch of tiny matrixes, each with one
segment (64 bytes) per row. This allows us to decode any matrix
individually; without this, random-access partial downloads are
not possible.

Unfortunately, this hurts performance significantly, but I was able
to ameliorate that somewhat by eliminating all allocations and
judiciously reusing buffers. (As a consequence of this, Encode now
takes the output shards as a parameter, instead of allocating and
returning them for you.)

Currently I'm seeing about 100-200 MB/s for a single thread, which
should be more than adequate. Sadly, some of these optimizations
require changing the guts of klauspost's reedsolomon package --
changes which I have very little expectation of him merging. So in
the next commit I'll be vendoring that package.

Lastly, I'd like to come back to this at some point and improve the
interop between the erasure coding and SectorBuilders. Specifically,
I want to encode directly into the SectorBuilder memory, instead of
encoding into buffers and then copying those buffers into the
SectorBuilders via Append. Performance concerns aside, I worry that
it's just too complicated to wire everything together properly, and
the SectorBuilders are at the crux of that, i.e. the intersection of
erasure-coding, encryption, and file-packing. In other words, the
goal should be to make it as easy as possible for some other dev to
implement renterutil.UploadDir from scratch.