renter: Remove shared subshards optimization
Since the compiler can now see the concrete type of the encoder,
it can determine that the subshards are not retained and can thus
be stack-allocated.

However, we still need to use a fancy trick: allocating a fixed-size
buffer and then immediately slicing it based on a variable. If we
naively allocated the buffer using a variable as its len, then the
compiler would still need to heap-allocate it, because it can't
determine its size at compile-time. But here, we know that the
erasure code can never use more than 256 shards, so we can fool
the compiler by statically allocating 256 and then trimming at
runtime as needed. (I used a similar trick within the vendored
reedsolomon package to prevent tons of allocations.)