user: Allow contract end height variants
Supporting these variants is uglier than I'd like, but I think they
will be appreciated. I typically think in terms of "form a contract
for 1000 blocks," not "for a contract until height 123456," and I
don't think I'm alone in that regard. The + syntax for extending
contracts is also nice to have, since when renewing, I think in
periods (e.g. renew every 1000 blocks) rather than in terms of
whatever the current block height is. Perhaps that's so common that
it should be made the default? Not sure.