user: Replace host whitelist with available/enabled
This is a pretty substantial overhaul. Previously, all contracts
were stored in a single folder, and the "hosts" array in the config
file selectively enabled or disabled them. This worked...okay, but
it was confusing and made migration especially unintuitive. In the
new design, we take an nginx-style approach: all the contracts you
form go in "contracts-available", and when you want to use one, you
"enable" it by symlinking to it in "contracts-enabled". This is nice
because it allows you to easily create sets of contracts with
different properties, like a set optimized for price vs. a set
optimized for speed. I also improved the ergonomics by adding two
commands for enabling+disabling contracts, so you can quickly
blacklist a host right from the terminal without having to open your
config file.

One thing I'm still uncertain about is "archived" contracts. The core
issue is, after you renew a contract, what do you do with the old
one? Strictly speaking, the "available" and "enabled" folders should
only contain a single contract per host. (This is necessary because
metafile don't specify an exact contract, just a host; so you need an
unambigous mapping from host->contract.) But in practice, I worry
that it's too easy to form two contracts with the same host, and end
up stuck in a confusing situation. I guess I can add more special
handling (to detect duplicates when forming), but it feels like hacks
piled on hacks. A more radical solution would be to not complain
about duplicates at all, and when you need to pick just one, use a
reasonable metric (such as "highest endheight"). But of course this
has many problems of its own; namely, if the metric selects the
"wrong" contract, how does the user figure out the problem and get
un-stuck? I'll have to think about it more.