user: Strip .usa suffix from served files
The server behavior now matches the behavior of the FUSE filesystem.
To be honest, I'm ambivalent about this change; the suffixes are an
eyesore, sure, but stripping them introduces the possibility of
collisions: if you have a folder called 'foo' and a metafile called
'foo.usa', what happens? As of this commit, metafiles take
precedence, so it's impossible to access the 'foo' folder. Ideally,
it would be impossible to create the colliding name in the first
place, but in practice we can't prevent that, since it happens at
the filesystem level. Oh well. This is a toy implementation anyway,
so I'm not terribly worried about covering all the bases.