renter: Make NewShardUploader key arg explicit
Previously, NewShardUploader took a hostIndex argument, which it
used internally to derive the encryption key. Before that, you
didn't pass anything at all: NewShardUploader used m.HostIndex to
get the index, and used that to derive the key. Unfortunately this
broke down because during migration, the new hosts aren't in the
metafile yet, so m.HostIndex can't find them. Passing an explicit
index fixed the problem, but it's not a very intuitive API. Using
the key directly instead makes the API more obvious, at the cost of
some extra code at the callsite.

Now I'm wondering if perhaps NewShardUploader should revert to its
original behavior and derive the key from the host's public key. The
Key field is exported, so the migration code would just set it
manually after creating the object. That feels dangerous...but it
could be worth it if it makes NewShardUploader easier to use in 90%
of cases.