renterutil: Fix download verification
The download verification code was assuming we used the old "large
shard" erasure-encoding, and thus was broken when we switched to
"small shard" encoding in 6565fd2d. The new algorithm runs the file
data through the actual encoder (using an m-of-m code, since we only
care about data shards) before verification. This is nicely generic;
it should continue working no matter how we change the erasure code.
(Not that I'm planning to do that anytime soon...)

This is the sort of bug that thorough testing would definitely catch.
It always bums me out a little to see bugs like this crop up weeks
after the breaking change was introduced. Have to do something about
that.