proto: Introduce ContractKey interface
Basically, we split the RenterKey field of ContractRevision out into its
own object, and incorporate it explicitly into ContractEditor (instead
accessing the key with Revision().SecretKey, which is ridiculous).

This paves the way for another change: making FormContract take a key
instead of generating one itself. The only downside of this is that it
takes a bit more work to form a contract: it's now the caller's
responsibility to generate the key, and they have to pass it explicitly
to SaveContract as well.