renter: Implement new shard format
Each SectorSlice now has a revision number, which we use as part of
the ChaCha nonce. This means we can securely modify any 64-byte
segment of a file without sacrificing the speed of ChaCha.

Also switch the SectorSlice checksum from Blake2b to CRC32. Now that
the host supplies Merkle proofs for partial downloads, we don't need
to verify them via checksum. However, keeping a non-cryptographic
checksum around allows us to verify contents on-disk, which is nice
when e.g. resuming a download.

Unfortunately this means we need to bump the metafile version number,
and possibly add conversion functions as well. At the moment I feel
too lazy to do that, so I'll wait to implement it until someone
complains. If you're really desperate, just download with an old
version and reupload.