renter: Use new renter-host protocol everywhere
This is unfortunately a bit messy. It may be prudent to restore the
old Upload and Sector methods, since they clarify a lot of code that
works exclusively with sectors.

I'm not attempting to do a proper job of integrating the new protocol
into the renter yet; that will come later, along with more dramatic
changes to PseudoFile/DownloaderSet. Honestly, I'm not even sure if
this code works with real hosts. I just want to strip out the old
renter-host protocol code.