ghost: Add package
ghost is barely-functional host, providing just enough functionality
to perform the renter-host protocol. It doesn't announce, submit
storage proofs, or watch the blockchain at all. It also does not
charge any fees or validate the revisions sent by the renter, though
I'll probably change that in order to more fully test that the renter
is behaving properly.

My goal is to eventually spin this out into a fully-functional host,
but for now I just need something minimal for testing purposes.