renterhost: Use separate read/write objBuffers
This prevents a data race that occurs when you try to read and
write concurrently. In practice, this should only happen during
the Read RPC, so it's kind of a bummer that we have to implement
a more general fix (especially since it doubles the max memory size
of a Session). On the other hand, it does make sense to have
separate buffers for reading and writing, since TCP is full duplex.