merkle: Implement Build/VerifyDiffProof
Not the most efficient implementation (definitely want to get rid of
those maps), but it works, which means proto.Session should now be
fully compatible with v1.4.0 hosts.

Verifying proofs is super slow, but nearly all of the cost comes from
calculating the SectorRoots of appended sectors. Yet another reason
to investigate optimizing that.