merkle: Unexport Stack
Stack was only exported so that it could be used in the contract
format, which is no longer necessary. Unexporting also lets us
remove some unused methods, in particular (Un)MarshalSia. The only
downside is that now we have a field called (stack).stack.

I also decided to change appendLeafHash to appendLeaf, since you can
now achieve appendLeafHash with insertNodeHash(h, 0).

The stack implementation is now pleasingly tiny: about 40 lines. Of
course, it's entirely possible that I may need to re-export it
someday. But doing so would be a trivial amount of work. Better to
adhere to YAGNI and keep things as minimal as they can be.