renter: Unify per-shard encryption keys
Now that we're using XChaCha20 with random nonces, there's no reason
to derive separate keys for each host; XChaCha20 derives a subkey
from the nonce and "key," meaning the "key" is really just a seed.

This allowed for a few simplifications. Since the encryption "key" is
always the same for each host, I was able to remove the EncryptionKey
method entirely; instead, callers reference the (MetaIndex).MasterKey
field directly. Consequently, the key argument was removed from
NewShardUploader, since it can be accessed from the m argument.

Overall, a definite win for simplicity. The only thing that makes me
nervous here is the unhindered access to key material within the
code. But I don't think this is a regression from the previous
approach; either way, if malicious code can access your metafiles,
it's game over.