renterutil: Move to FileSystem abstraction
Grouping file access via a filesystem makes sense, because then you
can serialize/prioritize/consolidate access to hosts. This is most
apparent when writing, because ideally you should be grouping the
writes of multiple metafiles into one new sector instead of uploading
a new sector for each 1 KB write.

This also means we can clean up a bunch of cruft in renterutil: the
separate upload, uploadDir, download, downloadDir, and downloadStream
are all made redundant by the filesystem abstraction, as long as it
properly consolidates reads and writes. The serve+mount code in user
is also simplified, essentially becoming just wrappers around the fs.

One open question is how migrations should be handled. Currently I'm
leaning towards incorporating them into the filesystem abstraction,
such that a migration becomes just one of many concurrent actions
that make use of the shared host sessions. But to be quite frank,
migration is an ugly business and I'm not exactly itching to turn my
attention to it.