renterutil: Support append-only PseudoFiles
Also make PseudoFile an interface, making it easier to split the
read-only and writable implementations.

This is the first step towards making PseudoFiles writable as well
as readable. Initially, it will only be possible to append at the
end of a file. We can then wrap the PseudoFile in a bufio.Writer to
achieve performance and efficiency comparable to the old upload
function.

The next stage is to support writes at arbitrary locations. This will
also require buffering for efficiency, but we'll need to manage the
buffering internally rather than using a bufio.Writer (because the
writes are not guaranteed to be sequential). At this point, the
PseudoFile abstraction should be good enough to use as a drop-in
replacement for *os.File.

Finally, we'll move the upload buffers out of the individual files
and into the PseudoFS. By sharing buffers between files, we can
transparently implement the sector packing performed by uploadDir.
At this point, we should be able to remove all non-filesystem code
in renterutil except for the migration functionality.

I'm hoping that I can find a clean way to fit migration into the
filesystem abstraction. One idea is to represent a migration as a
filesystem-to-filesystem copy operation, like copying a file from
a USB stick to your computer. Ideally, we end up with a nice design
that also makes use of the shared buffers to efficiently migrate
directories. (Directory migration currently does not preserve the
sector packing performed by uploadDir.) At this point, the filesystem
package would likely be split out of renterutil, leaving just the
client and scanning code, which truly are utilities. :)

Oh, one more consideration is metrics tracking. It seems like this
will be lost in the transition to the filesystem abstraction, but it
seems likely that it could be reintroduced at the filesystem level at
a later time. I'm sad to leave it in a broken state for now, but the
new abstractions are well worth the cost.