renterutil: Support 'unaligned' Writes
Unaligned writes are writes smaller than the file's minimum chunk
size, i.e. SegmentSize * MinShards. The Shard format only deals in
segments; it can't represent a slice of data 13 bytes long. This is
problematic if you want to support arbitrary-sized Writes at
arbitrary offsets.

I thought of two ways to address this. First, you could take what I
call the "SSD approach." When you write 13 bytes to an SSD page, it's
really reading the full page, copying your 13 bytes into it, and
writing the full page back. We can do the same thing: when you write
less than the minimum chunk size, we download the old chunk, copy the
new data into it, then upload the whole thing.

That seems really wasteful and slow, though, so I decided to pursue
an alternative approach: changing the Shard format to support smaller
writes. To do this, I added extra fields to the SectorSlice type that
specify which parts of the chunk are data and which parts are not. So
when you write your 13 bytes, you make a SectorSlice that says it
should be trimmed to 13 bytes, and you update the adjacent slices to
shave off their 13 bytes accordingly.

This is definitely more efficient, but implementation-wise, it was a
nightmare. It meant that all Shard-based code had to start respecting
these new fields, which got annoying real fast. In particular, it
made downloading really suck, because instead of a simple skipWriter,
we now needed a 'trimWriter' that would receive a bunch of padded
chunks and figure out how to extract the actual data.

After wrangling with that for a while and banging my head against all
the stupid "sub-SectorSlice-slice minShards modulo segSize padding"
math, I threw in the towel and went with the SSD approach. Yeah, it's
less efficient, but it's also way simpler, which means less bug-prone
and easier to change later. (Notice that even though it's simpler, it
still ain't *simple* -- there's just a lot of inherent complexity
here.)

Anyway, we have unaligned Writes now, and that's awesome. We're so
close to feature-parity with *os.File. It's possible that adding
arbitrary-offset Writes will interact poorly with the SSD approach,
but I think I'll at least be able to hack something together that
works. After that... optimization binge, my favorite. :)