renter: Remove SectorSlice.Checksum field
Checksums have always been gross, although the ability to resume
downloads is nice. While I was implementing Truncate, I noticed that
it can invalidate existing checksums (since it changes NumSegments),
and that was the straw that broke the camel's back -- not worth it.
I think we can get pretty much the same degree of functionality by
seeking to the offset of the last complete SectorSlice, and just
assuming that the contents are correct.

As a nice side effect, we can use proper 24-byte XChaCha nonces. 20
bytes would have been perfectly secure, but it's one less thing to
worry about.