user: Add 'optimistic' download resuming
It's "optimistic" because it may fail if the last write to the file
was interrupted, leaving garbage data. Without a checksum I don't
think it's possible to detect this, but we could do slightly better
than the "optimistic" version by seeking to the last SectorSlice-
aligned offset instead of the end of the file.